Fix incompatible poiner variable(32bit->64bit)

+++ lz/tornado/MatchFinder.cpp
@@ -7,7 +7,8 @@
 // If you need a commercial license to use the code, please write to Bulat.Ziganshin@gmail.com
 
 typedef uint   HashVal;      // Result of hashing function
-typedef uint32 PtrVal;       // Pointers to buf stored in HTable
+//typedef uint32 PtrVal;       // Pointers to buf stored in HTable
+typedef uint64 PtrVal;       // Pointers to buf stored in HTable
 typedef uint32 HintVal;      // Cached bytes from buf stored in HTable
